HUD
I created the HUD with a guitext object called score. Score needs to be
added to every scene for it to work. Score uses the GameManagerScript. I
didn't change how the score works, so it uses variables that were
already in other scripts, so it's a little funny with how score is
incremented. All script changes I made are only in the
GameManagerScript.